chore: clean up providers
